Migrate to building on for core20
This PR updates the snap build configuration. Previously with no base specified it was building against an older Ubuntu base of 16.04. By specifying core20 as the base, we get newer packages to build with, including the new GNOME extension which simplifies the parts, and makes the snap a little smaller and faster to start. The newer base also triggers newer code path in snapcraft.

I tested this on my local machine. All works as expected.

The only regression from doing this is that you'll no longer get i386 builds as i386 is no longer supported on Ubuntu 20.04 (the base from which core20 derives). If that's important we could go to core18 instead.

To build locally you must use `snapcraft` with `--enable-experimental-extensions`. This will land in `snapcraft` stable soon (end of the week I am told) so there will be no need for that experimental parameter, and will allow this to build in launchpad or the snapcraft build service. I'll keep this PR as draft until then.